% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_names_castles.R
\name{clean_names_castles}
\alias{clean_names_castles}
\title{Title}
\usage{
clean_names_castles(dataframe)
}
\arguments{
\item{dataframe}{a dataframe with castles name}
}
\value{
return a dataframe with clean castle names
}
\description{
Title
}
\examples{

df<-data.frame(Name = "Dunnottar", County="Aberdeenshire")
clean_names_castles(df)

}
