% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{locations}
\alias{locations}
\title{Locations of castles extracted from google earth}
\format{
A sf object with 219 rows and 3 variables
}
\usage{
data(locations)
}
\description{
A dataset containing the name and coordinates of castles in Scotland
}
\details{
\itemize{
\item Name. Name of the castle (not edited)
\item Description. Geographic features (points)
\item geometry. Geographic coordinates of the castles
}
}
\keyword{datasets}
