% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_castles.R
\name{merge_castles}
\alias{merge_castles}
\title{Merge castles locations, names and visits}
\usage{
merge_castles(df1, df2, img1, img2, img3, img4, shapefile)
}
\arguments{
\item{df1}{a dataframe with locations}

\item{df2}{a dataframe with visit status}

\item{img1}{Icon for non visited castles}

\item{img2}{Icon for visited castles}

\item{img3}{Icon for closed castles}

\item{img4}{Icon when using fill function in ggplot (same for all factors)}

\item{shapefile}{a polygon shapefile with Scotland coastline}
}
\value{
return a clean dataframe with coordinates, names and visits of castles
}
\description{
Merge castles locations, names and visits
}
\examples{
data(visits)
data(locations)
data(reg)

merge_castles(visits,locations,"im1","im2","im3","im4",reg)

}
