% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files_import.R
\name{read_shp_file}
\alias{read_shp_file}
\title{Import the different files}
\usage{
read_shp_file(coastline, out1, out2, shetland = FALSE)
}
\arguments{
\item{coastline}{a polygon shapefile}

\item{out.regions}{names of regions to exclude}

\item{klm.file}{klm files containing visited castles}

\item{colnames}{Column names to add in the new dataframe}
}
\value{
different dataframes
}
\description{
Import the different files
}
