% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reg}
\alias{reg}
\alias{locations}
\title{Vector shapefile reprsenting the county borders in Scotland}
\format{
A sf object with 35 rows and 3 variables
}
\usage{
data(reg)
}
\description{
A shapefile with poygons representing the different counties of Scotland
}
\details{
\itemize{
\item NAME. Name of the castle (not edited)
\item DESCRIPTIO. County (polygons)
\item geometry. Geographic coordinates of the castles
}
}
\keyword{datasets}
