% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScottishCastles-package.R
\docType{package}
\name{ScottishCastles-package}
\alias{ScottishCastles}
\alias{ScottishCastles-package}
\title{ScottishCastles: Map of Scottish castles}
\description{
Package which creates a map of Scottish castles that I have visited/not visited/wer closed. It has been an example for me to develop my first package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/auponchon/ScottishCastles}
  \item Report bugs at \url{https://github.com/auponchon/ScottishCastles/issues}
}

}
\author{
\strong{Maintainer}: Aurore Ponchon \email{aurore.ponchon@abdn.ac.uk} (\href{https://orcid.org/0000-0002-5126-6269}{ORCID})

}
\keyword{internal}
