% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{topo_df}
\alias{topo_df}
\title{A dataframe representing the topography of Scotland}
\format{
A datafrme with 42314 rows and 3 variables
}
\usage{
data(topo_df)
}
\description{
Dataframe extracted from a raster layer
}
\details{
\itemize{
\item value. Altitude (m)
\item x. Longitude (projected in m)
\item y. Latitude (pojected in m)
}
}
\keyword{datasets}
