% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{visits}
\alias{visits}
\title{Castles which have been visited or not visited (closed)}
\format{
A data frame with 157 rows and 4 variables
}
\usage{
data(visits)
}
\description{
A dataset containing the name, county and visit status of some castles in Scotland
}
\details{
\itemize{
\item Name. Name of the castle (not edited)
\item City. City in which castle is
\item County. County in which castle is
\item Visited. Whether the castle was visited (1), not visited yet (0) or closed
}
}
\keyword{datasets}
